Convert class TimepointAndNumberOfSpotsLabel to a JPanel extending class
