--- src/main/java/org/mastodon/app/ui/TimepointAndNumberOfSpotsLabel.java
@@ -1,44 +1,0 @@
-package org.mastodon.app.ui;
-
-import org.mastodon.mamut.model.Model;
-import org.mastodon.model.TimepointModel;
-
-import javax.swing.JLabel;
-import javax.swing.SwingUtilities;
-
-/**
- *  A label that displays (and updates according to the timepoint model) the current timepoint and the number of spots in the
- *  current timepoint.
- *  
- * @author Stefan Hahmann 
- */
-public class TimepointAndNumberOfSpotsLabel extends JLabel
-{
-	private final TimepointModel timepointModel;
-
-	private final Model model;
-
-	/**
-	 * Creates a new label that displays the current timepoint and the number of spots in the current timepoint.
-	 * Registers itself as a listener to the timepoint model.
-	 * 
-	 * @param timepointModel
-	 *            the timepoint model to listen to
-	 * @param model
-	 *            the model to get the number of spots from
-	 */
-	public TimepointAndNumberOfSpotsLabel( final TimepointModel timepointModel, final Model model )
-	{
-		this.timepointModel = timepointModel;
-		this.model = model;
-		this.timepointModel.listeners().add( this::updateTimepointAndNumberOfSpotsLabel );
-		updateTimepointAndNumberOfSpotsLabel();
-	}
-
-	private void updateTimepointAndNumberOfSpotsLabel()
-	{
-		SwingUtilities
-				.invokeLater( () -> setText( "timepoint: " + timepointModel.getTimepoint() + "   spots: "
-						+ model.getSpatioTemporalIndex().getSpatialIndex( timepointModel.getTimepoint() ).size() ) );
-	}
-}
